iiif: moved resource to separate file
